Add substance information trait for use with samples and inventory
