+++ pydatalab/schemas/startingmaterial.json
@@ -3,6 +3,49 @@
   "description": "A model for representing an experimental sample, based on the connection\nwith cheminventory.net, which mixes container-level and substance-level\ninformation.",
   "type": "object",
   "properties": {
+    "chemform": {
+      "title": "Chemform",
+      "example": [
+        "Na3P",
+        "Na<sub>3</sub>P",
+        "LiNiO2@C",
+        "Na3+xP",
+        "LiNi1/3Co0.1Mn0.1O2"
+      ],
+      "type": "string"
+    },
+    "smiles": {
+      "title": "Smiles",
+      "aliases": [
+        "SMILES",
+        "smiles_representation"
+      ],
+      "type": "string"
+    },
+    "inchi": {
+      "title": "Inchi",
+      "type": "string"
+    },
+    "inchi_key": {
+      "title": "Inchi Key",
+      "type": "string"
+    },
+    "GHS_codes": {
+      "title": "Ghs H-Codes",
+      "examples": [
+        "H224",
+        "H303, H316, H319"
+      ],
+      "type": "string"
+    },
+    "molar_mass": {
+      "title": "Molecular Weight",
+      "type": "number"
+    },
+    "CAS": {
+      "title": "Substance Cas",
+      "type": "string"
+    },
     "synthesis_constituents": {
       "title": "Synthesis Constituents",
       "default": [],
@@ -166,10 +209,6 @@
       "type": "string",
       "format": "date-time"
     },
-    "CAS": {
-      "title": "Substance Cas",
-      "type": "string"
-    },
     "chemical_purity": {
       "title": "Chemical Purity",
       "type": "string"
@@ -178,32 +217,12 @@
       "title": "Full %",
       "type": "string"
     },
-    "GHS_codes": {
-      "title": "Ghs H-Codes",
-      "examples": [
-        "H224",
-        "H303, H316, H319"
-      ],
-      "type": "string"
-    },
     "size": {
       "title": "Container Size",
       "type": "string"
     },
     "size_unit": {
       "title": "Unit",
-      "type": "string"
-    },
-    "chemform": {
-      "title": "Molecular Formula",
-      "type": "string"
-    },
-    "molar_mass": {
-      "title": "Molecular Weight",
-      "type": "number"
-    },
-    "smiles_representation": {
-      "title": "Smiles",
       "type": "string"
     },
     "supplier": {
