+++ pydatalab/src/pydatalab/models/traits.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from pydantic import BaseModel, Field, root_validator
+from pydantic import BaseModel, Field, root_validator, validator
 
 from pydatalab.models.blocks import DataBlockResponse
 from pydatalab.models.people import Group, Person
@@ -151,12 +151,48 @@ class HasSynthesisInfo(BaseModel):
         return values
 
 
-class HasChemInfo:
-    smile: str | None = Field(None)
+class HasSubstanceInfo(BaseModel):
+    """Trait mixin for models that have substance information."""
+
+    chemform: str | None = Field(
+        example=["Na3P", "Na<sub>3</sub>P", "LiNiO2@C", "Na3+xP", "LiNi1/3Co0.1Mn0.1O2"],
+    )
+    """A string representation of the chemical formula or composition associated with this sample.
+
+    The representation is relatively free-form; clients are expected parse and interpret HTML markup for subscripts
+    and accept unicode characters for greek letters.
+
+    """
+
+    smiles: str | None = Field(None, aliases=["SMILES", "smiles_representation"])
     """A SMILES string representation of the chemical structure associated with this sample."""
+
     inchi: str | None = Field(None)
-    """An InChI string representation of the chemical structure associated with this sample."""
+    """An International Chemical Identifier (InChI) string representation of chemicals/molecules associated with this sample."""
+
     inchi_key: str | None = Field(None)
-    """An InChI key representation of the chemical structure associated with this sample."""
-    """A unique key derived from the InChI string."""
-    chemform: str | None = Field(None)
+    """A unique key derived from the InChI."""
+
+    GHS_codes: str | None = Field(
+        alias="GHS H-codes",
+        examples=["H224", "H303, H316, H319"],
+    )
+    """A string describing any GHS hazard codes associated with this item. See https://pubchem.ncbi.nlm.nih.gov/ghs/ for code definitions."""
+
+    molar_mass: float | None = Field(alias="Molecular Weight")
+    """Mass per formula unit, in g/mol."""
+
+    CAS: str | None = Field(alias="Substance CAS")
+    """The CAS Registry Number for the substance described by this entry."""
+
+    @validator("molar_mass")
+    def add_molar_mass(cls, v, values):
+        from periodictable import formula
+
+        if v is None and values.get("chemform"):
+            try:
+                return formula(values.get("chemform")).mass
+            except Exception:
+                return None
+
+        return v
